requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage==6.3.3
awscli==1.24.0
flake8
python-dotenv>=0.5.1

torch~=1.11.0
torchvision~=0.12.0
setuptools==62.2.0
neptune-client==0.16.2
tqdm~=4.64.0
src~=0.1.0
numpy~=1.22.3
ignite~=1.1.0
matplotlib~=3.5.2